Add ParentGroup param to Resolve-RDCManGroup
Adding a parameter for the parent group allows recursive calls to include a full path to each server object.

These full paths (or $_.FullName, in the resulting object) are built from the FullName property of each parent object.

For example: "/example.com/Applications Servers/SERVER01"

This preserves the context for a given server entry, even without the full configuration (for example, when only a flattened list of servers is desired).