Update param names to reflect accepted types
In this case, "ParentGroup" is ambiguous, and could potentially refer to a parent group object.

The new name, "ParentGroupName" is more specific without being too verbose.

The above hopefully avoids confusing the Group parameter (an XML object containing a group of entries) with the parent group name (a string path of each preceding group name).